pyproject.toml
[project]
name = "sensenova-u1"
version = "0.1.0"
description = "SenseNova-U1: Unifying Multimodal Understanding and Generation with NEO-Unify Architecture."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11,<3.12"
authors = [
    { name = "OpenSenseNova" },
]
keywords = [
    "multimodal",
    "vision-language-model",
    "unified-model",
    "understanding-and-generation",
    "native-vlm",
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies.
#
# Versions here are pinned to match the internal reference environment
# (`miniconda3/envs/lmms_engine`, Python 3.11, CUDA 12.8)
# so that checkpoints trained there can be loaded and run bit-exactly. When
# bumping any of these, please re-verify inference on a reference checkpoint.
dependencies = [
    "torch==2.8.0",
    "torchvision==0.23.0",
    "transformers==4.57.1",
    "tokenizers==0.22.1",
    "accelerate==1.10.1",
    "huggingface-hub==0.36.2",
    "safetensors==0.6.2",
    "sentencepiece==0.2.1",
    "numpy==2.2.0",
    "pillow==12.0.0",
    "tqdm==4.67.1",
    "packaging==25.0",
    # HTTP client used by `sensenova_u1.prompt_enhance.adapters`. Version is left
    # open because only the high-level AsyncClient is used and the API has been
    # stable since 0.24.
    "httpx>=0.27,<1",
    "pre-commit>=4.5.1",
]

[project.optional-dependencies]
# Optional flash-attention acceleration. When absent, the model transparently
# falls back to torch SDPA; see src/sensenova_u1/models/neo_unify/modeling_qwen3.py.
#
# The reference env uses a cp311 / torch 2.8 wheel (`flash_attn 2.8.3`). PyPI
# does not host CUDA-specific wheels, so users will typically install this
# manually from a local .whl that matches their torch + Python combination, e.g.:
#
#   uv pip install /path/to/flash_attn-2.8.3+cu12torch28cxx11abitrue-cp311-cp311-*.whl
flash = [
    "flash-attn>=2.8,<3",
]

gguf = [
    "gguf>=0.10.0",
    "diffusers>=0.30.0",
]
dev = [
    "ruff",
    "pytest",
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/OpenSenseNova/SenseNova-U1"
Repository = "https://github.com/OpenSenseNova/SenseNova-U1"

[project.scripts]
sensenova-u1 = "sensenova_u1:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sensenova_u1"]

# -----------------------------------------------------------------------------
# uv configuration: pin torch / torchvision to the PyTorch cu128 index.
#
# This makes `uv sync` reproducibly install the CUDA 12.8 build of
# torch==2.8.0 (+cu128), matching the reference env exactly. If your NVIDIA
# driver is too old for CUDA 12.8, swap the URL below to e.g.
# `https://download.pytorch.org/whl/cu126` and adjust torch/torchvision
# versions accordingly.
# -----------------------------------------------------------------------------
[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [
    "src/sensenova_u1/models/neo_unify",
    # Vendored verbatim from SenseNova-Skills; keep diff-clean for easy uprev.
    "src/sensenova_u1/prompt_enhance/adapters",
    # Third-party submodules — keep diff-clean from upstream.
    "evaluation/easi/EASI",
    "evaluation/easi/lightllm-stack/LightLLM",
]

[tool.ruff.lint]
select = ["I"]
